refactor(ekb): restructure document_classification into service-specific packages
